fix angle fp24 from rad writer

+++ src/packet/p_fullness.hpp
@@ -7,7 +7,7 @@ struct packet_fullness : public packet_base {
     packet_fullness() : packet_base(packet_t_set_fullness) {}
 
     uint16_t snakeId = 0; // 3-4, int16, Snake id
-    float fullness = 0.0f; // 5-7, int24, value / 16777215 -> fam
+    uint8_t fullness = 0; // 5-7, int24, value / 16777215 -> fam
 
     constexpr size_t get_size() const noexcept { return 8; }
 };
